tried to connect with db
